Fix: Honor 'enabled' flag in notification event rules
The notify_event function was only checking if a channel (email, webhook, apprise)
was enabled, but not checking the 'enabled' flag that the UI uses to disable entire
notification events.

This caused notifications to be sent even when users had unchecked the notification
in the UI, as long as a channel was configured in notify.yaml.

Fixes issue #30 - users receiving unwanted notifications for events they disabled.